Inline isPrimitive for less bytes.
Previously we had this in order to handle boolean/number/string/null
nodes inside patch. Currently, booleans and nulls are filtered out
and numbers are converted to string directly inside `h'.

This means that whatever nodes patch receive, will always be either
a string, in which case we'll use document.createTextNode inside
`createElementFrom' or an object, in which case we'll proceed as
usual.

The other instance where isPrimitive was used before is inside merge,
so this logic has been moved there. The reason we check for primitive
types inside merge is to support simple models that are not an object,
but a primitive type like a number of a boolean.

It may be worthwhile reducing this check even further to
  typeof b !== "object" inside `merge'.

:fire: