Fix regression caused by element.textContent = node.
The previous behavior had a bug that if node was a
TextNode and oldNode was not, the TextNode would be
removed forever after the non TextNode was created.

This commit removes this optimization completely.

The idea was to change the element's text content if
the element was a TextNode instead of replacing it
with a new TextNode with just different textContet.

Unless you are alternating between two TextNodes,
and even if you are, this optimization offers too little.

TextNodes are also cheaper than other Nodes and you are
most likely alternating between a mix of elements and
TextNodes and not only TextNodes.