Enable embedding self-contained apps in views
This allows an app to declare that it wants to be rendered at
a specific position among siblings within it's root, by
declaring the element at that position to the renderer.

The primary motivation for this, is that it is possible to embed
self contained apps inside views, by rendering a "dummy" virtual
node with an oncreate handler that instantiates the embedded app.

Without this change, however, apps embedded like this will be
rendered at the *end* of the list of siblings, because the
renderer does not know any better.

*With* this change, the instantiation of the embedded app can
inform the renderer to render the app in the position of the dummy
vnode