Feature: Allow update to take a reducer.
Inside a thunk, update usually takes a partial state and
merges it with the previous state producing a new state.

This commit extends update to also take a function that
we will immediately call with the latest state and like
an action/reducer, allows you to update the state by
returning a partial state.

You'd want to use this feature when you need to update
the state inside an async process and you are not sure
whether the state that was passed to the action that
started the async process has become out-of-date or not.

A reducer is a function that receives the state and returns
a partial state (or a reducer :mindblow:).

  actions: {
    someAyncAction(state) {
      return update => {
        longTaskWithCallback(someData, () =>
          update(state => ({ value: state.value + 1 }))
        )
      }
    }
  }

This allows you to get the current & latest state inside the
callback of some async call.