Don't lock render before calling view function.
Basically any actions called inside the view function or inside "functions called inside the view" (e.g., Components) can invalidate the state, and when the view function returns, the state is left "out of sync" with the new vnode.

The idea is to compute the vnode, but before patching the DOM, determine if one or more actions were called (i.e., detect if a re-render was scheduled) and if that's the case return early (don't patch, since render will be called again soon anyway), otherwise proceed as usual.