Fix typing issue with Component<>
`Component` return type is `VNode<Attributes> | View<State, Actions>`. These
two aren't compatible with each other. Therefore the `Component` can only be a
function that returns a `VNode`. This has to do with the `h` only returning a
`VNode<Attributes>`.

Another fix could be to allow `h<Attributes()` to return `VNode<..> | View<..>`
but this will trickle down other issues with the typings.

This way there is a distinct way of creating a `VNode` from `Component` and
creating the `View(state, actions)` function typed as `View<TState, TActions>`