Implement Lazy (#777)
A lazy view is a part of the VDOM that is evaluated only if one or more specified state properties change (#721). This concept is borrowed from Elm's Html.lazy. This feature is not related to V1 lazy components.

Any function you wrap inside of the new `Lazy` function will re-render if the props passed in have changed. It works by creating a special object which the VDOM patching algorithm uses to compare the "old" node props with the "new" node props. If the props are the same (`===`), the old VNode is used without having to compute it again. When the VDOM patching sees that `oldNode === newNode`, it skips traversing that part of the DOM tree, potentially saving a lot of work!