Use rAF with fallback to setTimeout to defer rendering
Previously, we were using a microtask, but this creates an issue
when toggling an element with an event handler causing the event
to fire twice (unless we force Hyperapp to create a new element).
This is one of several reasons frameworks like React implement a
synthetic event system. Hyperapp V1 didn't have this bug, because
we were using setTimeout to defer rendering.

setTimeout defers callbacks too late, Promise.resolve too early,
requestAnimationFrame just right.

- Prevents firing events twice (without keys, or e.stopPropagation)
- Prevents inputs from going out of sync if typing too fast, or from
  flickering if one changes input's value, e.g., to prevent the user
  from entering numbers in it.
- Doesn't interfere with effects using requestAnimationFrame to run
  animations like setTimeout does.
- Falls back to setTimeout if requestAnimationFrame is not available.

One gotcha about using requestAnimationFrame is that rendering
will be effectively paused in when running in background tabs,
which is a browser measure to improve perf and battery life.

After considering all options, I decided to run with rAF until
I can find a better alternative.