Remove lib (@hyperapp/*) scoped packages -- the world is yours!
Now it's up to you to create your own effects and subscriptions.

The intention is to give you more freedom in exchange for some
initial user friendliness and make Hyperapp easier to maintain
in the long term.

We still encourage you to create reusable effects and subscriptions,
rather than ad-hoc solutions to solve specific problems.