Refactor while making a couple convenience types.
`Transition` and `StateWithEffects` (especially the latter) were created to make typing the return types of actions easier to do. The benefit they provide is best seen when working with actions that return more than just an updated state.

As a bonus they allowed the type definitions here to be further streamlined.